1. Now, session based user sign up, sign in, and sign out is fully available. 1-1. Try to use resourceful routing features of Rails, and Rails' helpers are used to wrap up versioned routing. 1-2. You can access to user sign in and sign up from the content's index page, which is used as a root page. 1-3. The sign in is session based, and you can sign out(logout) at the same root (index) page. The UI would be redered differently by user auth; which is guaranteed by existence of a session, especially an 'active session'. 2. Still there're plenty of matters relatd with evaluation(validating values) and contents management. Those are the very next steps to be done.
[Comment]
Using alias in 'routes.rb' was a good way to bind specific url with a template.(don't forget when it's vesioned an according prefix is required)
And also, make it clear that if you want to use resourceful routing, you need to designate the model you want; if necessary, allowing specific actions only.
(Lastly, about the structural changes) The authentication process is not under the control of user controller any longer.
Instead, there's a new model 'session', also managed by DB, and the session controller deals with it.
(of course, there are a few more facts like using 'concern' and so on, but for now, just keep it simple)

[GitHub]Closing related GitHub Issues
resolved - #7
'Fix sign up auth.'